requirements.txt
fastapi
pydantic
requests
beautifulsoup4
google-generativeai
uvicorn